fix lol

+++ src/modules/bloggers-platform/comments/application/use-cases/commands/delete-comment.usecase.ts
@@ -1,12 +1,13 @@
 import { CommandHandler, ICommandHandler } from "@nestjs/cqrs";
 import { Types } from "mongoose";
-import { NotFoundException } from "@nestjs/common";
+import { ForbiddenException, NotFoundException } from "@nestjs/common";
 import { CommentsRepository } from "../../../infrastructure/comments.repository";
 
 
 export class DeleteCommentCommand {
     constructor(
-        public readonly commentId: Types.ObjectId
+        public readonly commentId: Types.ObjectId,
+        public readonly userId: Types.ObjectId
     ) { }
 }
 
@@ -17,11 +18,15 @@ export class DeleteCommentUseCase
         private readonly CommentsRepository: CommentsRepository
     ) { }
 
-    async execute({ commentId }: DeleteCommentCommand): Promise<void> {
+    async execute({ commentId, userId }: DeleteCommentCommand): Promise<void> {
         const comment = await this.CommentsRepository.findCommentById(commentId)
 
         if (!comment) {
             throw new NotFoundException('Comment was not found')
+        }
+
+        if (comment.commentatorInfo.userId.toString() !== userId.toString()) {
+            throw new ForbiddenException()
         }
 
         await this.CommentsRepository.delete(comment)
